rename StyledLink fixture to kebab case
